+++ packages/vue/src/color-picker/color-picker-area.tsx
@@ -1,13 +1,12 @@
 import type { ColorAreaProps } from '@zag-js/color-picker'
 import { computed, defineComponent, type PropType } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import type { ComponentWithProps } from '../utils'
 import { ColorPickerAreaProvider } from './color-picker-area-context'
 import { useColorPickerContext } from './color-picker-context'
 
 export type ColorPickerAreaProps = HTMLArkProps<'div'> & ColorAreaProps
 
-export const ColorPickerArea: ComponentWithProps<ColorPickerAreaProps> = defineComponent({
+export const ColorPickerArea = defineComponent({
   name: 'ColorPickerArea',
   props: {
     xChannel: {
@@ -20,17 +19,15 @@ export const ColorPickerArea: ComponentWithProps<ColorPickerAreaProps> = defineC
     },
   },
   setup(props, { slots, attrs }) {
+    const api = useColorPickerContext()
     const areaProps = computed<ColorAreaProps>(() => ({
       xChannel: props.xChannel,
       yChannel: props.yChannel,
     }))
-
-    const rootContext = useColorPickerContext()
-
     ColorPickerAreaProvider(areaProps)
 
     return () => (
-      <ark.div {...rootContext.value.getAreaProps(areaProps.value)} {...attrs}>
+      <ark.div {...api.value.getAreaProps(areaProps.value)} {...attrs}>
         {slots.default?.()}
       </ark.div>
     )