+++ packages/vue/src/color-picker/color-picker-channel-slider-background.tsx
@@ -1,26 +1,20 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import type { ComponentWithProps } from '../utils'
 import { useColorPickerChannelSliderContext } from './color-picker-channel-slider-context'
 import { useColorPickerContext } from './color-picker-context'
 
 export type ColorPickerChannelSliderBackgroundProps = HTMLArkProps<'div'>
 
-export const ColorPickerChannelSliderBackground: ComponentWithProps<ColorPickerChannelSliderBackgroundProps> =
-  defineComponent({
-    name: 'ColorPickerChannelSliderBackground',
-    setup(_, { slots, attrs }) {
-      const sliderContext = useColorPickerChannelSliderContext()
+export const ColorPickerChannelSliderBackground = defineComponent({
+  name: 'ColorPickerChannelSliderBackground',
+  setup(_, { slots, attrs }) {
+    const api = useColorPickerContext()
+    const sliderProps = useColorPickerChannelSliderContext()
 
-      const rootContext = useColorPickerContext()
-
-      return () => (
-        <ark.div
-          {...rootContext.value.getChannelSliderBackgroundProps(sliderContext.value)}
-          {...attrs}
-        >
-          {slots.default?.()}
-        </ark.div>
-      )
-    },
-  })
+    return () => (
+      <ark.div {...api.value.getChannelSliderBackgroundProps(sliderProps.value)} {...attrs}>
+        {slots.default?.()}
+      </ark.div>
+    )
+  },
+})