+++ packages/vue/src/color-picker/color-picker-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/color-picker'
-import { type ComputedRef, type UnwrapRef } from 'vue'
 import { createContext } from '../context'
 import { type UseColorPickerReturn } from './use-color-picker'
 
-export const [ColorPickerProvider, useColorPickerContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('ColorPickerContext')
+export type ColorPickerContext = UseColorPickerReturn
 
-export type ColorPickerContext = UnwrapRef<UseColorPickerReturn>
+export const [ColorPickerProvider, useColorPickerContext] =
+  createContext<ColorPickerContext>('ColorPickerContext')