+++ packages/vue/src/color-picker/color-picker-eye-dropper-trigger.tsx
@@ -1,6 +1,5 @@
-import { defineComponent, h } from 'vue'
-import type { HTMLArkProps } from '../factory'
-import { useUniqueChild } from '../utils'
+import { defineComponent } from 'vue'
+import { ark, type HTMLArkProps } from '../factory'
 import { useColorPickerContext } from './color-picker-context'
 
 export type ColorPickerEyeDropperTriggerProps = HTMLArkProps<'button'>
@@ -9,10 +8,10 @@ export const ColorPickerEyeDropperTrigger = defineComponent({
   name: 'ColorPickerEyeDropperTrigger',
   setup(_, { slots, attrs }) {
     const api = useColorPickerContext()
-    return () => {
-      const DefaultSlot = useUniqueChild(slots, 'ColorPickerEyeDropperTrigger')
-
-      return h(DefaultSlot, { ...api.value.eyeDropperTriggerProps, ...attrs })
-    }
+    return () => (
+      <ark.button {...api.value.eyeDropperTriggerProps} {...attrs}>
+        {slots.default?.()}
+      </ark.button>
+    )
   },
 })