+++ packages/vue/src/color-picker/use-color-picker.ts
@@ -1,20 +1,33 @@
 import * as colorPicker from '@zag-js/color-picker'
-import { normalizeProps, useMachine } from '@zag-js/vue'
-import { computed, reactive, type ExtractPropTypes } from 'vue'
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
+import { computed, type ComputedRef } from 'vue'
+import { useEnvironmentContext } from '../environment'
+import type { Optional } from '../types'
 import { useId } from '../utils'
-import type { ColorPickerContext } from './color-picker'
 
-export const useColorPicker = <T extends ExtractPropTypes<ColorPickerContext>>(
+export type UseColorPickerProps = Optional<colorPicker.Context, 'id'> & {
+  modelValue?: colorPicker.Context['value']
+}
+export type UseColorPickerReturn = ComputedRef<colorPicker.Api<PropTypes>>
+
+export const useColorPicker = (
+  props: UseColorPickerProps,
   emit: CallableFunction,
-  context: T,
-) => {
-  const reactiveContext = reactive(context)
+): UseColorPickerReturn => {
+  const getRootNode = useEnvironmentContext()
+  const context = computed(() => {
+    const { modelValue, ...rest } = props
+    return {
+      ...rest,
+      value: modelValue,
+    }
+  })
 
   const [state, send] = useMachine(
     colorPicker.machine({
-      ...reactiveContext,
-      id: reactiveContext.id || useId().value,
-      value: reactiveContext.modelValue ?? reactiveContext.value,
+      ...context.value,
+      id: context.value.id ?? useId().value,
+      getRootNode,
       onChange(details) {
         emit('change', details)
         emit('update:modelValue', details.value)
@@ -23,9 +36,8 @@ export const useColorPicker = <T extends ExtractPropTypes<ColorPickerContext>>(
         emit('change-end', details)
       },
     }),
+    { context },
   )
 
   return computed(() => colorPicker.connect(state.value, send, normalizeProps))
 }
-
-export type UseColorPickerReturn = ReturnType<typeof colorPicker.connect>